+++ lab1/source/trivial.h
@@ -2,5 +2,5 @@
 typedef unsigned char Digit;
 typedef Digit* Number;
 
-extern int countNumberViaEnum(int n, int k);
+extern size_t countNumberViaEnum(size_t n, size_t k);
